+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/PhoneNumberController.kt
@@ -26,6 +26,8 @@ class PhoneNumberController internal constructor(
      */
     override val fieldValue: Flow<String> = _fieldValue
 
+    private val _hasFocus = MutableStateFlow(false)
+
     private val countryConfig = CountryConfig(
         overrideCountryCodes,
         tinyMode = true,
@@ -49,21 +51,39 @@ class PhoneNumberController internal constructor(
         )
     )
 
+    private val phoneNumberMinimumLength = countryDropdownController.selectedIndex.map {
+        PhoneNumberFormatter.lengthForCountry(
+            countryConfig.countries[it].code.value
+        )
+    }
+
     /**
      * Flow of the phone number in the E.164 format.
      */
     override val rawFieldValue = combine(fieldValue, phoneNumberFormatter) { value, formatter ->
         formatter.toE164Format(value)
     }
-    override val isComplete = fieldValue.map { it.isNotBlank() || showOptionalLabel }
+    override val isComplete = combine(fieldValue, phoneNumberMinimumLength) { value, minLength ->
+        value.length >= (minLength ?: 0) || showOptionalLabel
+    }
     override val formFieldValue = fieldValue.combine(isComplete) { fieldValue, isComplete ->
         FormFieldEntry(fieldValue, isComplete)
     }
 
-    override val error: Flow<FieldError?> = flowOf(null)
+    override val error: Flow<FieldError?> = combine(
+        fieldValue,
+        isComplete,
+        _hasFocus
+    ) { value, complete, hasFocus ->
+        if (value.isNotBlank() && !complete && !hasFocus) {
+            FieldError(R.string.incomplete_phone_number)
+        } else {
+            null
+        }
+    }
 
     internal val placeholder = phoneNumberFormatter.map { it.placeholder }
-    internal val prefix = phoneNumberFormatter.map { it.prefix }
+
     internal val visualTransformation = phoneNumberFormatter.map { it.visualTransformation }
 
     fun getCountryCode() = phoneNumberFormatter.value.countryCode
@@ -84,6 +104,10 @@ class PhoneNumberController internal constructor(
     override fun onRawValueChange(rawValue: String) {
         // any value can be treated the same way since it goes through clean up and formatting
         onValueChange(rawValue)
+    }
+
+    fun onFocusChange(newHasFocus: Boolean) {
+        _hasFocus.value = newHasFocus
     }
 
     companion object {