+++ paymentsheet-example/src/androidTestDebug/java/com/stripe/android/TestMultiStepFieldsReloaded.kt
@@ -170,6 +170,7 @@ class TestMultiStepFieldsReloaded {
                 paymentMethod = lpmRepository.fromCode("affirm")!!,
                 merchantCountryCode = "US",
                 currency = Currency.USD,
+                shipping = Shipping.On
             )
         )
     }